Create To Be Or Not To Be
Explanation
expect(val) Function: This function takes an initial value val and returns an object with toBe and notToBe methods.

toBe(expected) Method: Compares val with expected. If they are strictly equal (===), it returns {"value": true}. Otherwise, it throws an error with the message "Not Equal".

notToBe(expected) Method: Compares val with expected. If they are not strictly equal (!==), it returns {"value": true}. Otherwise, it throws an error with the message "Equal".